backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
mysql-connector-python==8.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyMySQL==1.1.0
marshmallow==3.20.1
python-dateutil==2.8.2 